Adding the ability for the bot to update people's channel name upon them changing their nickname
NOTE: This implementation does not require any additions to the database and uses information from discord to deduce which private channel belongs to which user.  This relies on a couple of assumptions (being that their is only one person who can see the private channel, other than admins, and that there are no public channels in the private channel category) which should hold up in the use cases this bot is being developed for.  However, note that this implementation may need to be updated in order to make the bot more robust for other use cases.